+++ Editor/SaintsEditorWindow.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-namespace SaintsField.Editor
-{
-    public class SaintsEditorWindow : MonoBehaviour
-    {
-
-    }
-}